requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
